security conf

+++ src/main/java/com/webtodolist/repository/ProjectRepository.java
@@ -7,10 +7,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.webtodolist.model.Project;
 import com.webtodolist.model.User;
 
-public interface ProjectRepository extends JpaRepository<Project, Long> {
+public interface ProjectRepository extends JpaRepository<Project, Integer> {
     List<Project> findByUser(User user);
     
-    List<Project> findByUserIdOrderByDataCreationProjectDesc(Long userId);
+    List<Project> findByUserIdOrderByDataCreationProjectDesc(int userId);
     
     List<Project> findByTitleContainingIgnoreCase(String title);
 }
